@@ -2,19 +2,19 @@ source 'https://rubygems.org'
 
 ruby '2.6.5'
 
-source "https://rails-assets.org" do
-  gem "rails-assets-bootstrap"
-  gem "rails-assets-bootstrap-material-design"
-  gem "rails-assets-knockout"
-  gem "rails-assets-arrive"
+source 'https://rails-assets.org' do
+  gem 'rails-assets-bootstrap', '~> 3'
+  gem 'rails-assets-bootstrap-material-design', '~> 0.3.0'
+  gem 'rails-assets-knockout'
+  gem 'rails-assets-arrive'
 end
 
 # Rails i18n
-gem 'rails-i18n', git: 'https://github.com/svenfuchs/rails-i18n.git', branch: 'rails-4-x'
+gem 'rails-i18n', git: 'https://github.com/svenfuchs/rails-i18n.git', branch: 'rails-5-x'
 # Use Puma as application server
 gem 'puma'
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.8'
+gem 'rails', '5.0'
 # Use SCSS for stylesheets
 gem 'sass-rails'
 # Use Uglifier as compressor for JavaScript assets
@@ -27,7 +27,7 @@ gem 'jquery-rails'
 # Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
 gem 'turbolinks'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder'
 
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
@@ -43,13 +43,13 @@ gem 'tzinfo-data'
 
 group :development, :test do
   # Use sqlite3 as the database for Active Record
-  gem 'sqlite3'
+  gem 'sqlite3', '~> 1.3.6'
 
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
 
   # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
+  gem 'web-console'
 
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
