@@ -1,7 +1,7 @@
-class Card < ActiveRecord::Base
+class Card < ApplicationRecord
   belongs_to :category
   belongs_to :cardset
-  
+
   validates :question, presence: true
   validates :answer, presence: true
   validates :category_id, presence: true
