@@ -1,4 +1,4 @@
-class Cardset < ActiveRecord::Base
+class Cardset < ApplicationRecord
   has_many :cards, :dependent => :destroy
 
   validates :title, presence: true
