@@ -1,9 +1,8 @@
-class Category < ActiveRecord::Base
-
+class Category < ApplicationRecord
   has_many :cards, :dependent => :destroy
 
   def self.colors
-    ['red', 'pink', 'purple', 'deep-purple', 'indigo', 'blue', 'light-blue', 'cyan', 'teal', 'green', 'light-green', 'lime', 'yellow', 'amber', 'orange', 'deep-orange', 'brown', 'grey', 'blue-grey']
+    %w(red pink purple deep-purple indigo blue light-blue cyan teal green light-green lime yellow amber orange deep-orange brown grey blue-grey)
   end
 
   validates :title, presence: true
